feat(tuning): per-part microtonal tuning — firmware raga restore, custom tables, Scala import
Section B+C of the plan (.pi/plans/2026-08-16-fixes-microtonal-scala-plan.md):

- TuningTables: the 30 firmware raga scale tables vendored verbatim
  from ambika_reference/controller/resources.cc (GPL-3.0, NOTICES
  updated), programmatically verified value-by-value; 12x int16
  offsets in 1/128-semitone units, 32767 = class-mute sentinel.
- part_raga APVTS choice param on PartData byte 4 (raga presets are
  hardware-faithful: the byte round-trips .PRO/.MUL and plays back
  on real Ambika hardware; firmware CC/NRPN row 46 is UI-only, so
  no CC/NRPN mapping).
- Custom per-part 12-entry tables: Part tuning state staged MT->AT
  via tuningDirty_ (existing dirty-flag discipline), pushed to
  voices, consumed at the single AmbikaVoice::startNote choke point
  (raw note % 12 — note-off/allocator keying untouched); AcceptNote
  gates refuse muted classes exactly like firmware part.cc.
- Engine state v7: length-prefixed per-part tuning block appended
  (FX-block pattern); restore accepts 1..7, v6 restores as 12-EDO,
  v8+ strict-rejected. .parvati multi gains guarded tuning_mode /
  tuning_offsets fields (voice_slots precedent; old files load as
  12-EDO, unknown keys ignored by old builds).
- Scala import (Source/ScalaImport.{h,cpp}): complete .scl/.kbm
  parser and converter, accurate to hardware limits — cents
  quantized to the 1/128-semitone grid (oscillator resolution is
  1/64 st), unmapped kbm keys become the 32767 mute sentinel,
  non-octave periods and >12-degree scales without a reducing .kbm
  are REJECTED, kbm root rotation and reference-pitch fold (with
  clamp + warning) honored, every outcome carries a warnings/errors
  list for the UI.
- Standing-bend pickup: per-channel wheel latch (defaults to centre
  8192) applied at voice trigger — voices no longer miss a latched
  wheel/MPE bend position.
- Tests: parvati_tuning_test (tables, hook mapping, sentinel gates,
  staging, v7 round-trip, .parvati/.MUL persistence, standing bend)
  and parvati_scala_import_test (grammar edges + expected-table
  corpus); descriptor-count tests updated for part_raga.

No Source/dsp behavior changes; bit-exactness policy intact.

+++ tests/roundtrip_test.cpp
@@ -106,9 +106,9 @@ int main()
         {
             char m[96];
             std::snprintf (m, sizeof (m), ".MUL part %d patch[112] byte-equal", i);
-            check (m1.parts[i].hasPatch == m2.parts[i].hasPatch && m1.parts[i].patch == m2.parts[i].patch, m);
+            check (m1.parts[(size_t) i].hasPatch == m2.parts[(size_t) i].hasPatch && m1.parts[(size_t) i].patch == m2.parts[(size_t) i].patch, m);
             std::snprintf (m, sizeof (m), ".MUL part %d part[84] byte-equal", i);
-            check (m1.parts[i].hasPart == m2.parts[i].hasPart && m1.parts[i].part == m2.parts[i].part, m);
+            check (m1.parts[(size_t) i].hasPart == m2.parts[(size_t) i].hasPart && m1.parts[(size_t) i].part == m2.parts[(size_t) i].part, m);
         }
         t.deleteFile();
     }
